batch_checkName_wcvp

Use the World Checklist of Vascular Plants (WCVP) database to check accepted names and update synonyms, in batch

batch_checkName_wcvp(
  occ = NA,
  wcvp_names = wcvp_names,
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard"
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

wcvp_names

WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/

if_author_fails_try_without_combinations

option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses

wcvp_selected_fields

WCVP fields selected as return, 'standard' basic columns, 'all' all available columns

Value

occ_checkName_wcvp: wcvp_plant_name_id, wcvp_taxon_rank, wcvp_taxon_status, wcvp_family, wcvp_taxon_name, wcvp_taxon_authors, wcvp_accepted_plant_name_id, wcvp_reviewed, wcvp_searchedName, wcvp_taxon_status_of_searchedName, wcvp_plant_name_id_of_searchedName, wcvp_taxon_authors_of_searchedName, wcvp_verified_author, wcvp_verified_speciesName, wcvp_searchNotes summary

Details

See help(checkName_wcvp) and https://powo.science.kew.org/about-wcvp

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# load package
library(ParsGBIF)

help(batch_checkName_wcvp)

wcvp_names <- get_wcvp(read_only_to_memory = TRUE)$wcvp_names
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#> [1] "downloading: http://sftp.kew.org/pub/data-repositories/WCVP//wcvp.zip"
#> [1] "loading: /tmp/RtmpOx6El6/wcvp_names.csv"
#> [1] " wcvp_names :1419265"
#> Warning: cannot remove file '/tmp/RtmpOx6El6/downlit', reason 'Directory not empty'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')
#> Loading required package: readr
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

res_batch_checkName_wcvp <- batch_checkName_wcvp(occ = occ,
                                                 wcvp_names =  wcvp_names,
                                                 if_author_fails_try_without_combinations = TRUE,
                                                 wcvp_selected_fields = 'standard')
#> Loading required package: dplry
#> Warning: there is no package called ‘dplry’
#> Loading required package: stringr
#> Loading required package: tidyselect
#> [1] "1-22 Achatocarpus nigricans Triana"
#> [1] "2-22 Achatocarpus obovatus Schinz & Autran"
#> [1] "3-22 Achatocarpus balansae Schinz & Autran"
#> [1] "4-22 Achatocarpus pubescens C.H.Wright"
#> [1] "5-22 Achatocarpus gracilis H.Walter"
#> [1] "6-22 Achatocarpus brevipedicellatus H.Walter"
#> [1] "7-22 Achatocarpus bicornutus Schinz & Autran"
#> [1] "8-22 Phaulothamnus spinescens A.Gray"
#> [1] "9-22 Achatocarpus microcarpus Schinz & Autran"
#> [1] "10-22 Achatocarpus spinulosus Griseb."
#> [1] "11-22 Achatocarpus praecox Griseb."
#> [1] "12-22 Achatocarpus praecox var. bicornutus (Schinz & Autran) Botta"
#> [1] "13-22 Achatocarpus praecox f. praecox"
#> [1] "14-22 Achatocarpus praecox f. obovatus (Schinz & Autran) Hauman"
#> [1] "15-22 Achatocarpus mexicanus H.Walter"
#> [1] "16-22 Achatocarpus hasslerianus Heimerl"
#> [1] "17-22 Achatocarpus praecox f. spinulosus (Schinz & Autran) Hauman"
#> [1] "18-22 Achatocarpus oaxacanus Standl."
#> [1] "19-22 Achatocarpus mollis H.Walter"
#> [1] "20-22 Achatocarpus brasiliensis H.Walter"
#> [1] "21-22 Achatocarpus praecox var. praecox"
#> [1] "22-22 Ampelocera hondurensis Donn.Sm."

names(res_batch_checkName_wcvp)
#> [1] "occ_checkName_wcvp" "summary"           

head(res_batch_checkName_wcvp$wcvpSummary)
#> NULL

head(res_batch_checkName_wcvp$wcvpOccurrence)
#> NULL