prepare_collectorsDictionary

Returns the list with the last name of the main collector associated with the unique key recordedBy.

prepare_collectorsDictionary(
  occ = NA,
  collectorDictionary_file =
    "https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv"
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

collectorDictionary_file

Collector dictionary file - point to a file on your local disk or upload via git at https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv.

Value

Ctrl_nameRecordedBy_Standard, Ctrl_recordedBy, Ctrl_notes, collectorDictionary, Ctrl_update, collectorName, Ctrl_fullName, Ctrl_fullNameII, CVStarrVirtualHerbarium_PersonDetails

Details

If recordedBy is present in the collector's dictionary, it returns the checked name, if not, it returns the last name of the main collector, extracted from the recordedBy field. If recordedBy is present in the collector's dictionary, returns the main collector's last name associated with the single recordedBy key, otherwise, returns the main collector's last name, extracted from the recordedBy field. It is recommended to curate the main collector's surname, automatically extracted from the recordedBy field. The objective is to standardize the last name of the main collector. That the primary botanical collector of a sample is always recognized by the same last name, standardized in capital letters and non-ascii characters replaced

See also

Examples

help(prepare_collectorsDictionary)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

collectorsDictionaryFromDataset <- prepare_collectorsDictionary(occ = occ,
                                                                collectorDictionary_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv')
#> [1] "Loading collectorDictionary..."
#> [1] "Extracting the main collector's surname...."
#> Warning: no non-missing arguments to max; returning -Inf

colnames(collectorsDictionaryFromDataset)
#> [1] "Ctrl_notes"                           
#> [2] "Ctrl_update"                          
#> [3] "Ctrl_nameRecordedBy_Standard"         
#> [4] "Ctrl_recordedBy"                      
#> [5] "collectorName"                        
#> [6] "Ctrl_fullName"                        
#> [7] "Ctrl_fullNameII"                      
#> [8] "CVStarrVirtualHerbarium_PersonDetails"
head(collectorsDictionaryFromDataset)
#>   Ctrl_notes Ctrl_update Ctrl_nameRecordedBy_Standard
#> 1       <NA>        <NA>                      ACEVEDO
#> 2       <NA>        <NA>                       ACOSTA
#> 3       <NA>        <NA>                       ACOSTA
#> 4       <NA>        <NA>                       ADARVE
#> 5       <NA>        <NA>                      AGUILAR
#> 6       <NA>        <NA>                      AGUILAR
#>                                   Ctrl_recordedBy collectorName Ctrl_fullName
#> 1                         R. ACEVEDO R.; C. REYES          <NA>          <NA>
#> 2                     R. ACOSTA P.; F. VÁZQUEZ B.          <NA>          <NA>
#> 3                      R. ACOSTA P.; N. ACOSTA B.          <NA>          <NA>
#> 4                                       J. ADARVE          <NA>          <NA>
#> 5                                   AGUILAR, R.M.          <NA>          <NA>
#> 6 JOSÉ AGUILAR CANO;SANDRA MEDINA;JHONATAN GUZMÁN          <NA>          <NA>
#>   Ctrl_fullNameII CVStarrVirtualHerbarium_PersonDetails
#> 1            <NA>                                  <NA>
#> 2            <NA>                                  <NA>
#> 3            <NA>                                  <NA>
#> 4            <NA>                                  <NA>
#> 5            <NA>                                  <NA>
#> 6            <NA>                                  <NA>

collectorDictionary_checked_file <- paste0(tempdir(),'/','collectorsDictionaryFromDataset.csv')

collectorDictionary_checked_file
#> [1] "/tmp/RtmpOx6El6/collectorsDictionaryFromDataset.csv"

write.csv(collectorsDictionaryFromDataset,
          collectorDictionary_checked_file,
          row.names = FALSE,
          fileEncoding = "UTF-8",
          na = "")