select_digital_voucher_and_sample_identification

Extract gbif issue

select_digital_voucher_and_sample_identification(
  occ = NA,
  occ_gbif_issue = NA,
  occ_checkName_wcvp = NA,
  occ_collectorsDictionary = NA,
  enumOccurrenceIssue = NA
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

occ_gbif_issue

= result of function extract_gbif_issue()$occ_gbif_issue

occ_checkName_wcvp

= result of function batch_checkName_wcvp()$occ_checkName_wcvp

occ_collectorsDictionary

= result of function update_collectorsDictionary()$occ_collectorsDictionary

enumOccurrenceIssue

An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)

Value

matchStatusDuplicates - "matched", "unmatched: no recordedBy and no recordNumber", "unmatched: no recordNumber" or "unmatched: no recordedBy" numberTaxonNamesSample - count of the different accepted scientific names, identified up to or below the specific level, listed in the sample duplicates, or Zero, if there is no identification, equal to or below the specific level, for the sample. sampleTaxonName - TAXON_name accepted and identified up to or below the specific level selected for the sample. sampleIdentificationStatus - 'Identified', 'divergent identifications', or 'unidentified'

Details

To group duplicates: 1) If the key to group duplicates is incomplete: Sample duplicates cannot be grouped due to missing collector information and/or collection number. Each record is considered a sample, with no duplicates. Select a voucher for each sample. Or, 2) If the key to group duplicates is complete: Group duplicates. Select a voucher, with the highest information score, among the duplicates in the sample. How to calculate the information score? moreInformativeRecord = sum of verbatim quality + quality of geospatial information. verbatim quality = sum of the number of flags with verbatim quality equal to TRUE. quality of geospatial information = If there is a geospatial issue, consider the one with the highest priority, with the highest score. To select sample taxonomic identification: 1) If the key for grouping duplicates is complete: select the accepted TAXON_NAME, identified up to or below the specific level, most frequent among the duplicates in the sample. If tie between frequency of accepted TAXON_NAME, identified up to or below the specific level: select the first accepted TAXON_NAME, identified up to or below the specific level, in alphabetical order. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample. Or, 2) If the key for grouping duplicates is incomplete: select TAXON_NAME, if accepted and identified up to or below the specified level. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample.

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

help(select_digital_voucher_and_sample_identification)

head(occ)
#> Error in eval(expr, envir, enclos): object 'occ' not found
head(res_gbif_issue$occ_gbif_issue)
#> Error in eval(expr, envir, enclos): object 'res_gbif_issue' not found
head(res_checkName_wcvp$occ_checkName_wcvp)
#> Error in eval(expr, envir, enclos): object 'res_checkName_wcvp' not found
head(res_collectorsDictionary$occ_collectorsDictionary)
#> Error in eval(expr, envir, enclos): object 'res_collectorsDictionary' not found
res_digital_voucher_and_sample_identification <- select_digital_voucher_and_sample_identification(occ = occ,
                                                                                                  occ_gbif_issue = res_gbif_issue$occ_gbif_issue,
                                                                                                  occ_checkName_wcvp = res_checkName_wcvp$occ_checkName_wcvp,
                                                                                                  occ_collectorsDictionary = res_collectorsDictionary$occ_collectorsDictionary,
                                                                                                  enumOccurrenceIssue = EnumOccurrenceIssue)
#> Error in if (is.na(enumOccurrenceIssue)) {    data(EnumOccurrenceIssue)} else {    EnumOccurrenceIssue <- enumOccurrenceIssue}: the condition has length > 1

names(res_digital_voucher_and_sample_identification)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found

head(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found
colnames(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found

head(res_digital_voucher_and_sample_identification$occ_join_results)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found
colnames(res_digital_voucher_and_sample_identification$occ_join_results)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found